+++ changelog.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## 2023-01-17
+
+### Changed
+- add more status code
+- change stock snack_fk relation to CASCADE
+
+
 ## 2023-01-16
 
 ### Changed